Extract federation config logic to its own file
